refactor(suite): fix ReduxModalProps type-only circular dep
